fix: address bootstrapping issue
While modifying the scripts I introduced an issue where if the checksum
bucket was empty no new processing would occur. This change updates the
scripts to always compute the current interval insuring that at least
new intervals are computed. As time passes this will eventually cover
all intervals.

When outer join lands as a feature in Flux we can change this script to
be more robust against older out of date intervals.